requirements.txt
pip==25.1.1
requests==2.32.4
pytest==8.4.0
python-dotenv==1.1.0